2009-02-12	Joel Sherrill <joel.sherrill@oarcorp.com>
	* cpu.c, rtems/score/cpu.h: Change prototype of IDLE thread to
	consistently return void * and take a uintptr_t argument.

+++ cpukit/score/cpu/m32c/cpu.c
@@ -147,7 +147,7 @@ void _CPU_Install_interrupt_stack( void )
  *  XXX document implementation including references if appropriate
  */
 
-void *_CPU_Thread_Idle_body( uint32_t ignored )
+void *_CPU_Thread_Idle_body( uintptr_t ignored )
 {
 
   for( ; ; )
